Remove a circular dependency in the evaluators
Since this is an recursive descent parser, there is actually a circular
dependency by design. There's not really much one can do with that fact.
One could define every evaluator in the same file and exploit hoisting,
but that isn't particularly maintainable. Instead, we propogate the
reference to the root evaluator townwards and use that reference when we
need to eveluate a new expression.